+++ src/PathfindingSystem.h
@@ -12,7 +12,7 @@ public:
     std::pair<float, float> getNextPathPoint(const EntityManager& entityManager, int entityID);
 
 private:
-    std::vector<std::pair<int, int>> findPath(const std::vector<std::vector<int>>& grid, Point start, Point goal);
+    std::vector<std::pair<float, float>> findPath(const std::vector<std::vector<int>>& grid, Point start, Point goal);
 };
 
 #endif
